apps/ai-service/requirements.txt
# FastAPI and ASGI Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18

# Google Gemini AI
google-generativeai==0.8.3

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client
httpx==0.28.1
aiohttp==3.11.10
requests==2.32.3

# Audio processing (for future use with HuggingFace)
numpy==2.1.3
scipy==1.14.1

# Utilities
python-dotenv==1.0.1
